python/homework10/requirements.txt
python-dotenv==0.21.0
python-telegram-bot==13.15
pytest~=7.2.0